fix swiper last-child issue
When I tried the swiper style in my own app, I didn't need the
last-child style setting width:100%. Upon further investigation, I
realized this was needed in the codeshow app because of the
margin-left:120px. The margin pushes the flexbox off the screen and the
last-child width:100% was an attempt to regain what was lost. Resetting
the width on the last child did not fix all this issues, hence the
#swipeparent {width:600px} in flexbox.css. The real issue is best seen
via the horizontal scrollbar (may only be visual on a non-touch device).
The h-scroll does not have right arrow to scroll incrementally because
it is pushed off the screen. All this is say, simply changing the width
of the swiper from 100% to auto fixes all (I hope) the issues
alleviating the need for the last-child style.

You will notice I made some other minor changes associated with using
the swiper class.

~Thanks for all these examples! I'm learning a lot!