fix render fail for demos with sections on subsequent loads
When a user went to a demo with multiple sections, it was working the first time and then
failing every subsequent time. I talked with the WinJS test team and Sean Lee helped me
discover that my code for rendering sections should return a promise which will make the
demo page wait on it before being "ready".

I also changed the order of the demos in D3 to show the most impressive one first.