improve code viewing in demo view
Code viewing is not entirely finished. For demos that have sections,when the user chooses to view the code they should see the demo for the section they were viewing. Likewise, it would be ideal if they navigated to another code set and then back to the demo, they should be scrolled to the correct demo